src/server/motionDetectorDjangoServer/requirements.txt
absl-py==2.3.1
asgiref==3.9.1
astunparse==1.6.3
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.3.3
cycler==0.12.1
Django==5.2.5
django-cors-headers==4.9.0
django-extensions==4.1
django-rest-knox==5.0.2
djangorestframework==3.16.1
flatbuffers==25.2.10
fonttools==4.59.2
gast==0.6.0
google-pasta==0.2.0
grpcio==1.74.0
gunicorn==23.0.0
h5py==3.14.0
idna==3.10
keras==3.11.3
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
ml_dtypes==0.5.3
namex==0.1.0
numpy==2.2.6
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pillow==11.3.0
protobuf==6.32.0
psycopg2-binary==2.9.10
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.4
rich==14.1.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
tensorflow-hub==0.16.1
termcolor==3.1.0
tf_keras==2.20.1
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
